add coverage
When running the command, if there are any non-covered lines in the current view
a tiny circle with a scope name "error" shows up next to uncovered lines.

When the view changes, the inline circles will dissapear.
There is also a "go_remove_coverage" command that will only show up when the current
view has any non-covered lines.

Closes #5